pyproject.toml
[project]
name = "thesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "kaggle>=1.8.3",
    "kagglehub>=0.4.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "nvidia-ml-py>=13.590.48",
    "opencv-python>=4.13.0.90",
    "pillow>=12.1.0",
    "pynvml>=13.0.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.54.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "ultralytics>=8.4.9",
]
